Move movie details feature to its own package
