+++ MovieCatalog/Packages/3-Features/MovieDetails/Sources/MovieDetails/Internal/MovieDetailsDataSource.swift
@@ -1,0 +1,23 @@
+import SwiftUI
+import MovieModels
+import MovieComponents
+import MovieDBNetworking
+
+struct MovieDetailsDataSource: Sendable {
+    var movieDetails: @Sendable (MovieID) async throws -> MovieDetailsViewData
+
+    static let live: MovieDetailsDataSource = .init(
+        movieDetails: { movieID in
+            let dto = try await MovieDBClient().movieDetails(for: movieID)
+            return MovieDetailsViewData(dto: dto)
+        }
+    )
+
+    static func previewClient() -> MovieDetailsDataSource {
+        .init(movieDetails: { _ in .previewValue() })
+    }
+}
+
+extension EnvironmentValues {
+    @Entry var movieDetailsDataSource: MovieDetailsDataSource = .live
+}
